+++ packages/logging/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: 'Logging facility.',
-  version: '2.0.0'
+  version: '2.0.0-beta200.6'
 });
 
 Npm.depends({